requirements.txt
appnope==0.1.2
backcall==0.1.0
certifi==2019.9.11
chardet==3.0.4
decorator==4.4.0
Django==2.2.5
django-filter==2.4.0
djangorestframework==3.12.4
gunicorn==19.9.0
idna==2.8
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
Markdown==3.3.4
numpy==1.20.2
pandas==1.2.4
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==2.0.9
psycopg2-binary==2.8.6
ptyprocess==0.6.0
Pygments==2.4.2
python-dateutil==2.8.1
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
urllib3==1.25.6
wcwidth==0.1.7
